About to totally restructure the game History.
Will convert from 3 vars into a currentTurnNumber, and a history Array that holds a status Object with 3 keys in each of its elements.
This will enable adding to the UI a turn number display as well as Undo AND Redo buttons that switch to MUI "disabled" when necesary.
add turn number to UI display