Elevated all Map Making functions and Game Constants to App.js
Whe the MapMaking functions and Game Constants lived inside the <ClassicGame> it was calling the map making functions everytime the Game had to re render, which was every time that its state was updated.  By elevating these functions that Only need to be called ONCE from the <ClassicGame> into the <App> and then passing the results to the <Game> as props I prevent them from being called on every rerender of the game.